JMK_122: package refactoring
