+++ src/main/resources/application.properties
@@ -18,5 +18,5 @@ spring.kafka.listener.ack-mode=manual
 
 aws.region=ap-south-1
 
-ses.sender.email=pawanpreetsingh193@gmail.com
+ses.sender.email=
 server.port=8081
